Am realizat doua module: 1. Module - pentru clasele utiolizate ulterior in Swiming Competition - contine o clasa Entity abstracta care va da ID-ul fiecarei clase derivate 2. Repository - pentru a gestiona obiectele din Module - o interfata generala, pentru fiecare intervata ulterioara cu metodele generale - interfete pentru fiecare obiect, pentru a introduce noi metode ce vor fi ulterior implementate Tot proiectul este legat folosind gradle: in settings.gradle se introduc modulele create pentru a putea naviga intre ele obiecte/interfete etc. Ambele module pot fi rulate, insa nu acest lucru se doreste,  urmand sa introduc un nou modul care imi permite rularea intregului proiect.
Nou adaugate:
1. Implementarea Repository-lor care utilizeaza baze de date SQLite si fisiere de configurare 'db.config'.
2. Partea de jurnalizare folosind un fisier de tip xml. Tot ce reiese din logger va fi salvat in fisierul app.log.
3. Mici modificari la nivel de Model (Entity) pentru a folosi incrementarea automata a ID-ului si a pastra unicitatatea acestuia in tabelele baze de date (primary key).

Nou adaugate:
1. Service - interfete de service si implementarile lor
2. Controller - o clasa pentru logarea celor din oficiu, una pentru gestionarea de inregistrare si cautare participanti si un Application unde lansam efectiv aplicatia in executie.

+++ Model/src/main/java/org/example/Event.java
@@ -1,6 +1,6 @@
 package org.example;
 
-public class Event extends Entity<Integer> {
+public class Event extends Entity<Long> {
     private int distance;
     private String style;
     //private List<Participant>  participants;
